refactor(ECO-3080): Separate dexscreener.ts into multiple files
